Move capture sources to capture
