+++ projects/optic/src/commands/capture/interactions/documented.ts
@@ -6,7 +6,7 @@ import { generateEndpointSpecPatches } from '../patches/patches';
 
 import { updateSpecFiles } from '../../oas/diffing/document';
 import { SpecPatch } from '../../oas/specs';
-import { CapturedInteractions } from '../../oas/captures';
+import { CapturedInteractions } from '../sources/captured-interactions';
 import { ApiCoverageCounter } from '../coverage/api-coverage';
 import * as AT from '../../oas/lib/async-tools';
