+++ projects/optic/src/commands/capture/patches/patches.ts
@@ -4,7 +4,7 @@ import { checkOpenAPIVersion } from '@useoptic/openapi-io';
 
 import { ApiCoverageCounter } from '../coverage/api-coverage';
 import { OpenAPIV3, SpecPatch, SpecPatches } from '../../oas/specs';
-import { CapturedInteractions } from '../../oas/captures';
+import { CapturedInteractions } from '../sources/captured-interactions';
 import { DocumentedInteraction, Operation } from '../../oas/operations';
 import { DocumentedBodies } from '../../oas/shapes';
 import { createMissingMethodPatch } from '../../oas/specs/patches/generators/missing-method';
