+++ projects/optic/src/commands/capture/sources/__tests__/interaction.test.ts
