+++ projects/optic/src/commands/capture/sources/proxy.ts
@@ -6,8 +6,8 @@ import {
   CompletedBody,
   TimingEvents,
 } from 'mockttp';
-// import { getCA, CAOptions } from 'mockttp/dist/util/tls';
-import { Subject } from '../../../lib/async-tools';
+
+import { Subject } from '../../oas/lib/async-tools';
 import { pki, md } from 'node-forge';
 import { randomBytes } from 'crypto';
 import { Readable } from 'stream';
@@ -24,7 +24,7 @@ import globalLog from 'log';
 import invariant from 'ts-invariant';
 import chalk from 'chalk';
 import { UserError } from '@useoptic/openapi-utilities';
-import { logger } from '../../../../../logger';
+import { logger } from '../../../logger';
 type Logger = typeof globalLog;
 
 export const log: Logger = globalLog.get('captures:streams:sources:proxy'); // export so it can be enabled in testing
