+++ projects/optic/src/commands/capture/storage.ts
@@ -2,7 +2,8 @@ import crypto from 'crypto';
 import fs from 'node:fs/promises';
 import os from 'os';
 import path from 'path';
-import { CapturedInteraction, HttpArchive } from '../oas/captures';
+import { CapturedInteraction } from './sources/captured-interactions';
+import { HttpArchive } from './sources/har';
 import { OperationQueries } from '../oas/operations/queries';
 import { jsonPointerHelpers } from '@useoptic/json-pointer-helpers';
 import { getEndpointId } from '../../utils/id';
