+++ projects/optic/src/commands/oas/captures/index.ts
@@ -1,16 +1,1 @@
-import { CapturedBody } from '../../capture/interactions/body';
-
-export { CapturedBody };
-export { CapturedInteractions } from '../../capture/interactions/captured-interactions';
-
-export { CapturedInteraction } from '../../capture/interactions/captured-interactions';
-export type {
-  CapturedRequest,
-  CapturedResponse,
-} from '../../capture/interactions/captured-interactions';
-
-export { HarEntries } from './streams/sources/har';
-export type { HttpArchive } from './streams/sources/har';
-
-export { ProxyInteractions, ProxyCertAuthority } from './streams/sources/proxy';
 export { getInteractions } from './getInteractions';
