+++ projects/optic/src/commands/oas/diffing/patch.ts
@@ -17,7 +17,6 @@ import { jsonPointerHelpers } from '@useoptic/json-pointer-helpers';
 import chalk from 'chalk';
 import { ShapeDiffResult } from '../shapes/diffs';
 import { OpenAPIV3 } from '@useoptic/openapi-utilities';
-import { CapturedInteraction, CapturedInteractions } from '../captures';
 import { Subject, tap } from '../lib/async-tools';
 import { DocumentedInteraction, DocumentedInteractions } from '../operations';
 import { DocumentedBodies, DocumentedBody } from '../shapes';
@@ -25,6 +24,10 @@ import { updateReporter } from '../reporters/update';
 import { ApiCoverageCounter } from '../../capture/coverage/api-coverage';
 import { ParsedOperation } from './document';
 import { nextCommand } from '../reporters/next-command';
+import {
+  CapturedInteraction,
+  CapturedInteractions,
+} from '../../capture/sources/captured-interactions';
 
 export async function patchOperationsAsNeeded(
   patchInteractions: CapturedInteractions,
