+++ projects/optic/src/commands/oas/shapes/body.ts
@@ -1,8 +1,8 @@
 import { BodyLocation } from '@useoptic/openapi-utilities';
-import { CapturedBody } from '../../capture/interactions/body';
 import { ShapePatch } from '.';
 import { SchemaObject, Schema } from './schema';
 import { Option } from 'ts-results';
+import { CapturedBody } from '../../capture/sources/body';
 
 export interface Body {
   contentType?: string;
