+++ projects/optic/src/commands/oas/shapes/streams/documented-bodies.ts
@@ -10,12 +10,12 @@ import MIMEType from 'whatwg-mimetype';
 import { BodyExampleFacts, ComponentSchemaExampleFacts } from '../../specs';
 import { OpenAPIV3 } from '../../specs';
 import { Body, DocumentedBody } from '../body';
-import { CapturedBody } from '../../captures';
 import {
   DocumentedInteraction,
   findResponse,
   findBody,
 } from '../../operations';
+import { CapturedBody } from '../../../capture/sources/body';
 
 export type { DocumentedBody };
